Split out extraction tests, add a pinned env and a no-extras env.
