+++ tox.ini
@@ -15,6 +15,24 @@ commands =
         --cov=zyte_test_websites \
         {posargs:zyte_test_websites tests}
 
+[testenv:pinned]
+basepython = python3.9
+deps =
+    {[testenv]deps}
+    aiohttp==3.9.0
+    aiohttp_jinja2==1.6
+    cssselect==1.2.0
+    parsel==1.5.0
+    web-poet==0.14.0
+    zyte-common-items==0.24.0
+
+[testenv:no-extras]
+extras =
+commands =
+    python -m pytest \
+        --ignore-glob='tests/test_*_extraction.py' \
+        {posargs:zyte_test_websites tests}
+
 [testenv:pre-commit]
 deps =
     pre-commit
